# src/controllers/anomalies.controller.ts
import { Request, Response } from 'express';
import prisma from '../config/database.js';
import { AnomaliesResponse } from '../types/user.types.js';

export class AnomaliesController {
  static async getAnomalies(req: Request, res: Response): Promise<void> {
    const { severity = 'all', type = 'all', hours = '24' } = req.query;
    const startTime = Date.now();
    const apiKeyHash = (req as any).apiKeyHash;

    try {
      const response: AnomaliesResponse = {
        period: `last ${hours} hours`,
        total_anomalies: 0,
        anomalies: [],
        summary: {
          blocked_automatically: 0,
          flagged_for_review: 0,
          clean_signups: 0
        }
      };

      res.json({
        success: true,
        data: response,
        meta: {
          request_id: req.headers['x-request-id'] || 'anonymous',
          version: '1.0.0',
          processing_ms: Date.now() - startTime,
          from_cache: false
        }
      });
    } catch (error) {
      res.status(500).json({
        success: false,
        error: {
          code: 'internal_error',
          message: 'Failed to fetch anomalies',
          docs_url: 'https://docs.trustiq.io/errors'
        },
        meta: {
          request_id: req.headers['x-request-id'] || 'anonymous',
          version: '1.0.0',
          processing_ms: Date.now() - startTime,
          from_cache: false
        }
      });
    }
  }
}